Refactor project structure and organize dependencies
