+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "national-parks",
-  "version": "0.10.2",
+  "version": "0.10.3",
   "description": "United States National Parks",
   "main": "index.html",
   "license": "MIT",
@@ -20,20 +20,23 @@
     "angular": "^1.6.1",
     "angular-lazy-image": "^0.3.4",
     "angular-ui-router": "^0.4.2",
+    "body-parser": "~1.15.2",
     "bootstrap": "^3.3.7",
+    "cookie-parser": "~1.4.3",
+    "dotenv": "^4.0.0",
+    "express": "~4.14.0",
+    "flickr-sdk": "^2.0.1",
     "jquery": "^3.1.1",
+    "morgan": "~1.7.0",
+    "node-foursquare": "^0.3.0",
+    "request": "^2.79.0",
     "serve-favicon": "~2.3.0"
   },
   "devDependencies": {
     "babel-cli": "^6.24.0",
     "babel-preset-es2015": "^6.24.0",
-    "body-parser": "~1.15.2",
-    "cookie-parser": "~1.4.3",
     "debug": "~2.2.0",
     "del": "^2.2.2",
-    "dotenv": "^4.0.0",
-    "express": "~4.14.0",
-    "flickr-sdk": "^2.0.1",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-cache": "^0.4.6",
@@ -46,9 +49,6 @@
     "gulp-uglify": "^2.1.0",
     "gulp-usemin": "^0.3.28",
     "jshint": "^2.9.4",
-    "jshint-stylish": "^2.2.1",
-    "morgan": "~1.7.0",
-    "node-foursquare": "^0.3.0",
-    "request": "^2.79.0"
+    "jshint-stylish": "^2.2.1"
   }
 }
